LUN-310: Fix and move test
