+++ src/services/__tests__/AsyncStorage.spec.ts
@@ -1,8 +1,10 @@
 import RNAsyncStorage from '@react-native-async-storage/async-storage'
 
-import { ExerciseKeys, Progress } from '../../constants/data'
+import { ExerciseKeys, Progress, SIMPLE_RESULTS } from '../../constants/data'
+import { DocumentResult } from '../../navigation/NavigationTypes'
+import DocumentBuilder from '../../testing/DocumentBuilder'
 import { mockDisciplines } from '../../testing/mockDiscipline'
-import AsyncStorage from '../AsyncStorage'
+import AsyncStorage, { saveExerciseProgress } from '../AsyncStorage'
 
 describe('AsyncStorage', () => {
   const customDisciplines = ['first', 'second', 'third']
@@ -75,6 +77,25 @@ describe('AsyncStorage', () => {
     it('should not save progress for done exercise without improvement', async () => {
       await AsyncStorage.setExerciseProgress(1, ExerciseKeys.wordChoiceExercise, 0.5)
       await AsyncStorage.setExerciseProgress(1, ExerciseKeys.wordChoiceExercise, 0.4)
+      const progress = await AsyncStorage.getExerciseProgress()
+      expect(progress[1]).toStrictEqual({ [ExerciseKeys.wordChoiceExercise]: 0.5 })
+    })
+
+    it('should calculate and save exercise progress correctly', async () => {
+      const documents = new DocumentBuilder(2).build()
+      const documentsWithResults: DocumentResult[] = [
+        {
+          document: documents[0],
+          result: SIMPLE_RESULTS.correct,
+          numberOfTries: 1
+        },
+        {
+          document: documents[0],
+          result: SIMPLE_RESULTS.incorrect,
+          numberOfTries: 3
+        }
+      ]
+      await saveExerciseProgress(1, 1, documentsWithResults)
       const progress = await AsyncStorage.getExerciseProgress()
       expect(progress[1]).toStrictEqual({ [ExerciseKeys.wordChoiceExercise]: 0.5 })
     })
